tests(fixes) overall section: completed
