+++ src/__tests__/background/requests/operations/ops/create-claimed-account.test.ts
@@ -2,7 +2,6 @@ import { broadcastCreateClaimedAccount } from '@background/requests/operations/o
 import createClaimedAccount from 'src/__tests__/background/requests/operations/ops/mocks/create-claimed-account';
 import messages from 'src/__tests__/background/requests/operations/ops/mocks/messages';
 import userData from 'src/__tests__/utils-for-testing/data/user-data';
-
 describe('create-claimed-account tests:\n', () => {
   const { methods, constants, mocks } = createClaimedAccount;
   const { requestHandler, data, confirmed } = constants;
