+++ src/__tests__/background/requests/operations/ops/mocks/authority.ts
@@ -1,6 +1,8 @@
 import { RequestsHandler } from '@background/requests';
 import { ExtendedAccount, TransactionConfirmation } from '@hiveio/dhive';
 import {
+  KeychainKeyTypes,
+  KeychainRequestTypes,
   RequestAddAccountAuthority,
   RequestAddKeyAuthority,
   RequestId,
@@ -8,25 +10,49 @@ import {
   RequestRemoveKeyAuthority,
 } from '@interfaces/keychain.interface';
 import AccountUtils from 'src/utils/account.utils';
-import keychainRequest from 'src/__tests__/utils-for-testing/data/keychain-request';
+import mk from 'src/__tests__/utils-for-testing/data/mk';
+import userData from 'src/__tests__/utils-for-testing/data/user-data';
 import mocksImplementation from 'src/__tests__/utils-for-testing/implementations/implementations';
 
+const commonValues = {
+  domain: 'domain',
+  username: mk.user.one,
+};
+
 const requestHandler = new RequestsHandler();
+
 const data = {
   addAccountAuthority: {
-    ...keychainRequest.wValues.addAccountAuthority,
+    ...commonValues,
+    type: KeychainRequestTypes.addAccountAuthority,
+    role: KeychainKeyTypes.posting,
+    weight: 1,
+    authorizedUsername: 'theghost1980',
     request_id: 1,
   } as RequestAddAccountAuthority & RequestId,
   removeAccountAuthority: {
-    ...keychainRequest.wValues.removeAccountAuthority,
+    ...commonValues,
+    type: KeychainRequestTypes.removeAccountAuthority,
+    authorizedUsername: 'theghost1980',
+    role: KeychainKeyTypes.posting,
+    method: KeychainKeyTypes.active,
     request_id: 1,
   } as RequestRemoveAccountAuthority & RequestId,
   addKeyAuthority: {
-    ...keychainRequest.wValues.addKeyAuthority,
+    ...commonValues,
+    type: KeychainRequestTypes.addKeyAuthority,
+    authorizedKey: userData.one.encryptKeys.posting,
+    method: KeychainKeyTypes.active,
+    weight: 1,
+    role: KeychainKeyTypes.posting,
     request_id: 1,
   } as RequestAddKeyAuthority & RequestId,
   removeKeyAuthority: {
-    ...keychainRequest.wValues.removeKeyAuthority,
+    ...commonValues,
+    type: KeychainRequestTypes.removeKeyAuthority,
+    authorizedKey: userData.one.encryptKeys.posting,
+    method: KeychainKeyTypes.active,
+    role: KeychainKeyTypes.posting,
     request_id: 1,
   } as RequestRemoveKeyAuthority & RequestId,
 };
