+++ src/__tests__/background/requests/operations/ops/mocks/broadcast.ts
@@ -1,12 +1,22 @@
 import { RequestsHandler } from '@background/requests';
 import { ExtendedAccount, TransactionConfirmation } from '@hiveio/dhive';
-import { RequestBroadcast, RequestId } from '@interfaces/keychain.interface';
-import keychainRequest from 'src/__tests__/utils-for-testing/data/keychain-request';
+import {
+  KeychainKeyTypes,
+  KeychainRequestTypes,
+  RequestBroadcast,
+  RequestId,
+} from '@interfaces/keychain.interface';
+import mk from 'src/__tests__/utils-for-testing/data/mk';
 import mocksImplementation from 'src/__tests__/utils-for-testing/implementations/implementations';
 
 const requestHandler = new RequestsHandler();
+
 const data = {
-  ...keychainRequest.wValues.broadcastOperation,
+  domain: 'domain',
+  username: mk.user.one,
+  type: KeychainRequestTypes.broadcast,
+  operations: 'transfer',
+  method: KeychainKeyTypes.posting,
   request_id: 1,
 } as RequestBroadcast & RequestId;
 const confirmed = {
