+++ src/__tests__/background/requests/operations/ops/mocks/decode-memo.ts
@@ -1,14 +1,25 @@
 import { RequestsHandler } from '@background/requests';
 import { TransactionConfirmation } from '@hiveio/dhive';
-import { RequestDecode, RequestId } from '@interfaces/keychain.interface';
-import keychainRequest from 'src/__tests__/utils-for-testing/data/keychain-request';
+import {
+  KeychainKeyTypes,
+  KeychainRequestTypes,
+  RequestDecode,
+  RequestId,
+} from '@interfaces/keychain.interface';
+import mk from 'src/__tests__/utils-for-testing/data/mk';
 import mocksImplementation from 'src/__tests__/utils-for-testing/implementations/implementations';
 
 const requestHandler = new RequestsHandler();
+
 const data = {
-  ...keychainRequest.wValues.decode,
+  domain: 'domain',
+  username: mk.user.one,
+  type: KeychainRequestTypes.decode,
+  message: '',
+  method: KeychainKeyTypes.active,
   request_id: 1,
 } as RequestDecode & RequestId;
+
 const confirmed = {
   id: '1',
   trx_num: 112234,
